+++ CHANGELOG.md
@@ -23,6 +23,7 @@ while other changes are minor it's still an important update.
 
 ### Added
 - "storage list" command to neofs-lens (#2852)
+- "morph mint-balance" command to neofs-adm (#2867)
 
 ### Fixed
 - GETRANGE request may fail in certain cases (#2849)